feat: add docs addon and create routes constants
