+++ packages/design-system/.storybook/addons.js
@@ -1,2 +1,3 @@
 import '@storybook/addon-knobs/register';
 import '@storybook/addon-a11y/register';
+import '@storybook/addon-docs/register';
