+++ packages/design-system/.storybook/config.js
@@ -1,14 +1,19 @@
 import { configure, addParameters, addDecorator } from '@storybook/react';
 import { themes } from '@storybook/theming';
 import { withA11y } from '@storybook/addon-a11y';
+import { DocsPage, DocsContainer } from '@storybook/addon-docs/blocks'
 
-const req = require.context('../src', true, /\.story\.(ts|tsx)$/);
+const req = require.context('../src', true, /\.story\.(ts|tsx|mdx)$/);
 
 addDecorator(withA11y);
 
 addParameters({
   options: {
     theme: themes.dark,
+  },
+  docs: {
+    container: DocsContainer,
+    page: DocsPage,
   },
 });
