+++ packages/design-system/.storybook/presets.js
@@ -1,0 +1,1 @@
+module.exports = ['@storybook/addon-docs/react/preset'];
