+++ packages/sdk/src/products/product.provider.ts
@@ -3,12 +3,7 @@ import { http as httpClient } from '../utils/httpClient';
 import {
   TGetProducts,
 } from './product.interface';
-
-const ROOT_PRODUCTS = '/products';
-
-export const ENDPOINTS = {
-  PRODUCTS: `${ROOT_PRODUCTS}`,
-};
+import { ENDPOINTS } from '../utils/routes.constants';
 
 export interface IProductProvider {
   getProducts: TGetProducts;
