+++ packages/sdk/src/utils/routes.constants.ts
@@ -1,0 +1,9 @@
+export const api = {
+  BASE_URL: 'https://api.github.com/users/viniarruda/repos',
+};
+
+const ROOT_PRODUCTS = '/products';
+
+export const ENDPOINTS = {
+  PRODUCTS: `${ROOT_PRODUCTS}`,
+};
